Move targets to run RTEMS/FreeRTOS out of run_qemu.py
Keeping them with the target that builds the kernel makes more sense.
