+++ pycheribuild/projects/cross/freertos.py
@@ -30,9 +30,11 @@
 # SUCH DAMAGE.
 #
 import os
+from typing import ClassVar
 
 from .crosscompileproject import (CheriConfig, CompilationTargets, CrossCompileAutotoolsProject, DefaultInstallDir,
                                   GitRepository)
+from ..run_qemu import LaunchQEMUBase
 from ...config.loader import ComputedDefaultValue
 
 
@@ -59,6 +61,9 @@ class BuildFreeRTOS(CrossCompileAutotoolsProject):
 
     default_demo = "RISC-V-Generic"
     default_demo_app = "main_blinky"
+    demo: "ClassVar[str]"
+    demo_app: "ClassVar[str]"
+    demo_bsp: "ClassVar[str]"
 
     def __init__(self, config: CheriConfig):
         super().__init__(config)
@@ -143,7 +148,6 @@ class BuildFreeRTOS(CrossCompileAutotoolsProject):
             self.real_install_root_dir / str("FreeRTOS/Demo/" + self.demo + "_" + self.demo_app + ".elf"))
 
     def process(self):
-
         if self.demo not in self.supported_freertos_demos:
             self.fatal("Demo " + self.demo + "is not supported")
 
@@ -154,3 +158,52 @@ class BuildFreeRTOS(CrossCompileAutotoolsProject):
                           # Add compiler-rt location to the search path
                           LDFLAGS="-L" + str(self.compiler_resource / "lib")):
             super().process()
+
+
+class LaunchFreeRTOSQEMU(LaunchQEMUBase):
+    target = "run-freertos"
+    dependencies = ["freertos"]
+    supported_architectures = [CompilationTargets.BAREMETAL_NEWLIB_RISCV64_PURECAP,
+                               CompilationTargets.BAREMETAL_NEWLIB_RISCV64]
+    forward_ssh_port = False
+    qemu_user_networking = False
+    _enable_smbfs_support = False
+    _add_virtio_rng = False
+
+    default_demo = "RISC-V-Generic"
+    default_demo_app = "main_blinky"
+
+    @classmethod
+    def setup_config_options(cls, **kwargs):
+        super().setup_config_options(defaultSshPort=None, **kwargs)
+
+        cls.demo = cls.add_config_option(
+            "demo", metavar="DEMO", show_help=True,
+            default=cls.default_demo,
+            help="The FreeRTOS Demo to run.")  # type: str
+
+        cls.demo_app = cls.add_config_option(
+            "prog", metavar="PROG", show_help=True,
+            default=cls.default_demo_app,
+            help="The FreeRTOS program to run.")  # type: str
+
+        cls.demo_bsp = cls.add_config_option(
+            "bsp", metavar="BSP", show_help=True,
+            default=ComputedDefaultValue(function=lambda _, p: p.default_demo_bsp(),
+                                         as_string="target-dependent default"),
+            help="The FreeRTOS BSP to run. This is only valid for the "
+                 "paramterized RISC-V-Generic. The BSP option chooses "
+                 "platform, RISC-V arch and RISC-V abi in the "
+                 "$platform-$arch-$abi format. See RISC-V-Generic/README for more details")
+
+    def default_demo_bsp(self):
+        return "qemu_virt-" + self.target_info.get_riscv_arch_string(self.crosscompile_target, softfloat=True) + "-" + \
+               self.target_info.get_riscv_abi(self.crosscompile_target, softfloat=True)
+
+    def get_riscv_bios_args(self) -> "list[str]":
+        # Run a FreeRTOS demo application (run in machine mode using the -bios QEMU argument)
+        return ["-bios", str(BuildFreeRTOS.get_install_dir(self)) + "/FreeRTOS/Demo/" +
+                self.demo + "_" + self.demo_app + ".elf"]
+
+    def process(self):
+        super().process()
